Initial Anti Spoofing For Question View Counter
Originally if a question was viewed and the page refreshed then the view counter
would increase on each refresh. Now this has been averted.

A session side value is held for whether a user has viewed a particular question.
If not then the view is incremented and the session key set.

Obviously this isn't perfect, as deletion of session key can affect this.
More measures should be put in place.