Adds Initial Likes/Dislikes Functionality
Now the buttons for answers will update the likes or dislikes count
of the answer, and the overal user simulatenously.

Note that there is no checking, and spamming the buttons can affect
the scores. There is no visual feedback via button colour as discussed
yet. And there is now serverside longterm or midterm storage
of who likes/dislikes what, this is still to come.

Also need a check to ensure likes/dislikes for user/answer
can never be negative. Now I sleep...