Implements User Feedback On Like/Dislike
Now the buttons will update their colour based upon
whether a user likes or dislikes a certain answer.

Upon viewing a question, a user can immediately see what
votes they cast on the answers, this updates asynchronysly
via AJAX when they make a change.

Images are a bit WIP, and it was late, so might need some checks
on the changes of this commit. Couldn't find any major faults.